+++ R/recipe_commit.R
@@ -49,9 +49,8 @@ recipe_commit <- function(workdir, config) {
     DBI::dbDisconnect(con)
   })
 
-  success <- DBI::dbWriteTable(con, tbl, dat, append = TRUE)
-
   report_data_import(con, workdir, config)
+  success <- DBI::dbWriteTable(con, tbl, dat, append = TRUE)
 
   if (success) {
     ## I should do something here if the unlink fails, though I don't
@@ -123,12 +122,7 @@ report_read_data <- function(workdir, config) {
                         jsonlite::unbox(artefacts[[i]][[1]]$description)
   }
 
-  published_yml <- path_orderly_published_yml(workdir)
-  if (file.exists(published_yml)) {
-    published <- yaml_read(published_yml)$published
-  } else {
-    published <- FALSE
-  }
+  published <- report_is_published(workdir)
 
   if (!is.null(info$depends)) {
     used <- unique(vcapply(info$depends, "[[", "id"))
@@ -189,4 +183,10 @@ report_read_data <- function(workdir, config) {
     ret <- cbind(ret, as.data.frame(custom, stringsAsFactors = FALSE))
   }
   ret
+}
+
+
+report_is_published <- function(workdir) {
+  published_yml <- path_orderly_published_yml(workdir)
+  file.exists(published_yml) && yaml_read(published_yml)$published
 }